fix:手動でターン終了
