+++ frontend/src/lib/constants.ts
@@ -1,0 +1,9 @@
+// ゲーム関連の定数定義
+
+export const GAME_CONSTANTS = {
+    MAX_HP: 20,
+    MAX_PP: 10,
+    INITIAL_HP: 20,
+    INITIAL_PP: 0,
+    PP_PER_TURN: 1,
+};
