+++ frontend/src/lib/types.ts
@@ -1,0 +1,66 @@
+export interface MockUser {
+    id: string;
+    name: string;
+    isAdmin: boolean;
+}
+
+export interface MockRoomPlayer {
+    id: string;
+    roomId: string;
+    userId: string;
+    hp: number;
+    pp: number;
+    turn: number;
+}
+
+export interface MockRoom {
+    id: string;
+    ownerId: string;
+    status: 'waiting' | 'playing' | 'finish';
+    players: MockRoomPlayer[];
+    owner: MockUser;
+}
+
+// API関連の型定義
+export interface CreateRoomInput {
+    roomId?: string;
+    currentUser: MockUser;
+}
+
+export interface JoinRoomInput {
+    roomId: string;
+    currentUser: MockUser;
+}
+
+export interface GetRoomInput {
+    roomId: string;
+}
+
+export interface UpdatePlayerStatusInput {
+    roomId: string;
+    currentUser: MockUser;
+    hp?: number;
+    pp?: number;
+    turn?: number;
+}
+
+export interface StartTurnInput {
+    roomId: string;
+    currentUser: MockUser;
+}
+
+export interface EndTurnInput {
+    roomId: string;
+    currentUser: MockUser;
+}
+
+export interface ConsumePPInput {
+    roomId: string;
+    currentUser: MockUser;
+    ppCost: number;
+}
+
+export interface ForceEndOpponentTurnInput {
+    roomId: string;
+    currentUser: MockUser;
+}
